Added

+++ Testing/TestMod1_samQuiz.java
@@ -2,7 +2,7 @@ package Testing;
 
 public class TestMod1_samQuiz {
     public static void main(String[] args) {
-        int[] x = new int[] {1,2,3};
+        double[] x = new double[] {1,2,3};
         System.out.println(x[0]);
